+++ src/js/archi-rendering/archi-rendering.js
@@ -1,4 +1,3 @@
 import '../components/header/header';
 import './archi-rendering-section/archi-rendering-templates';
-import './archi-rendering-section/make-bold-first';
 import '../components/footer';
