+++ src/pages/seven-benefits-of-bim.html
@@ -23,7 +23,7 @@
     ></include>
     <include src="./partials/components/whatProblems.html"></include>
     <include
-      src="./partials/components/sevenBenefitsOfBim/disadvantagesSevenBenefits.html"
+      src="./partials/components/sevenBenefitsOfBim/disadvantages.html"
     ></include>
     <include
       src="./partials/components/sevenBenefitsOfBim/whereToFind.html"
