+++ src/partials/components/sevenBenefitsOfBim/disadvantages.html
@@ -2,7 +2,7 @@
   <div class="container">
     <h3 class="disadvantages__title">Disadvantages of BIM outsourcing</h3>
     <img
-      class="author__img"
+      class="disadvantages__img"
       src="../images/7-benefits/disadvantages/disadvantages@1x.jpg"
       alt=""
     />
@@ -11,9 +11,9 @@
       industry, the technology isn't perfect. Here's what some people consider
       to be its primary drawbacks:
     </p>
-    <ol>
+    <ul>
       <li>
-        <h3>01 Demanding technology</h3>
+        <h3 class="disadvantages__title">01 Demanding technology</h3>
         <p class="disadvantages__text">
           BIM is not just 3D modeling software. BIM technology is a set of
           software tools and methodologies, which are used to model graphical
@@ -25,13 +25,13 @@
           development phase. Some of the most popular modeling tools include
           expensive premium software such as:
         </p>
-        <ul>
-          <li>Autodesk Revit</li>
-          <li>Navisworks</li>
-          <li>Dynamo</li>
-          <li>SYNCHRO Pro</li>
-          <li>Solibri</li>
-          <li>Tekla</li>
+        <ul class="disadvantages__text">
+          <li class="disadvantages__softItems">Autodesk Revit</li>
+          <li class="disadvantages__softItems">Navisworks</li>
+          <li class="disadvantages__softItems">Dynamo</li>
+          <li class="disadvantages__softItems">SYNCHRO Pro</li>
+          <li class="disadvantages__softItems">Solibri</li>
+          <li class="disadvantages__softItems">Tekla</li>
         </ul>
         <p class="disadvantages__text">
           Companies wanting to embrace BIM should plan to invest resources to
@@ -45,7 +45,7 @@
         </p>
       </li>
       <li>
-        <h3>02 Reliant on cooperation</h3>
+        <h3 class="disadvantages__title">02 Reliant on cooperation</h3>
         <p class="disadvantages__text">
           Using BIM services requires active engagement and regular input from
           stakeholders. All parties have to be willing to invest their time and
@@ -63,7 +63,7 @@
         </p>
       </li>
       <li>
-        <h3>03 Lack of experts</h3>
+        <h3 class="disadvantages__title">03 Lack of experts</h3>
         <p class="disadvantages__text">
           Creating an accurate BIM model requires in-depth technical skills with
           various modeling software. You also need knowledgeable engineering and
@@ -76,6 +76,6 @@
           construction model is technically feasible.
         </p>
       </li>
-    </ol>
+    </ul>
   </div>
 </section>
