+++ src/partials/components/sevenBenefitsOfBim/whereToFind.html
@@ -1,41 +1,41 @@
-<section>
+<section class="section whereToFind">
   <div class="container">
-    <h2>Where to find BIM experts?</h2>
+    <h2 class="whereToFind__title">Where to find BIM experts?</h2>
     <img
-      class="author__img"
+      class="whereToFind__img"
       src="../images/7-benefits/where-to-find/where-to-find@1x.jpg"
       alt=""
     />
-    <p>
+    <p class="whereToFind__text">
       When people say BIM is not necessary for their project, they mean that
       it’s too complex for them. However, the right company can do the modeling
       for you.
     </p>
-    <p>
+    <p class="whereToFind__text">
       But finding the right BIM company for your construction project can be
       difficult. To make the most out of BIM technology, you need to consider
       the following:
     </p>
     <ul>
-      <li>
+      <li class="whereToFind__text__list">
         Engineering qualifications. Experts with an engineering background will
         understand your designs and be able to transform your flat drawings into
         fully realized 3D models.
       </li>
-      <li>
+      <li class="whereToFind__text__list">
         Value for the money. Some companies may charge high prices, but you will
         spend less overall if design problems are identified and fixed as early
         as possible.
       </li>
-      <li>
+      <li class="whereToFind__text__list">
         Experience in design. The right provider will know how to resolve your
         problems and help you stay within your budget.
       </li>
-      <li>
+      <li class="whereToFind__text__list">
         Legislative knowledge. Companies with legal experts on their team can
         ensure that your project adheres to local and governmental standards.
       </li>
-      <li>
+      <li class="whereToFind__text__list">
         Track record. A company with relevant case studies is much more likely
         to help you improve your project.
       </li>
