--- src/sass/components/_disadvantagesSevenBenefits.scss
@@ -1,2 +1,0 @@
-.disadvantages {
-}
