+++ src/sass/components/sevenBenefitsOfBim/whereToFind.scss
@@ -1,0 +1,35 @@
+.whereToFind {
+  padding-bottom: 20px;
+}
+
+.whereToFind__title {
+  @include font-title-20;
+  padding-bottom: 20px;
+}
+
+.whereToFind__img {
+  min-width: 320px;
+  padding-bottom: 40px;
+}
+
+.whereToFind__text,
+.whereToFind__text__list {
+  @include font-body-reg;
+  padding-bottom: 20px;
+}
+
+.whereToFind__text__list {
+  padding-left: 20px;
+  position: relative;
+}
+
+.whereToFind__text__list::before {
+  position: absolute;
+  top: 4px;
+  left: 0;
+  //   margin-right: 10px;
+  content: '';
+  width: 10px;
+  height: 9px;
+  background-image: url('../images/vectors/arrow-right-black.svg');
+}
