+++ src/sass/index.scss
@@ -57,7 +57,8 @@
 
 @import './components/whatProblems';
 
-@import './components/disadvantagesSevenBenefits';
+@import './components/sevenBenefitsOfBim/disadvantages';
+@import './components/sevenBenefitsOfBim/whereToFind.scss';
 
 @import './components/footer';
