+++ src/sass/utils/_fonts.scss
@@ -13,6 +13,14 @@
   line-height: calc(22 / 20);
 }
 
+@mixin font-title-20 {
+  font-family: 'PT Mono';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 100%;
+}
+
 @mixin font-title-14 {
   font-family: 'PT Mono';
   font-style: normal;
